# faster brand check