--- a/lib/web/fetch/formdata.js
+++ b/lib/web/fetch/formdata.js
@@ -256,6 +256,6 @@ function makeEntry (name, value, filename) {
   return { name, value }
 }
 
-webidl.is.FormData = webidl.util.MakeTypeAssertion(FormData.prototype)
+webidl.is.FormData = webidl.util.MakeTypeAssertion(FormData)
 
 module.exports = { FormData, makeEntry, setFormDataState }