--- a/lib/web/fetch/request.js
+++ b/lib/web/fetch/request.js
@@ -986,7 +986,7 @@ Object.defineProperties(Request.prototype, {
   }
 })
 
-webidl.is.Request = webidl.util.MakeTypeAssertion(Request.prototype)
+webidl.is.Request = webidl.util.MakeTypeAssertion(Request)
 
 // https://fetch.spec.whatwg.org/#requestinfo
 webidl.converters.RequestInfo = function (V, prefix, argument) {