--- a/lib/web/fetch/response.js
+++ b/lib/web/fetch/response.js
@@ -619,7 +619,7 @@ webidl.converters.ResponseInit = webidl.dictionaryConverter([
   }
 ])
 
-webidl.is.Response = webidl.util.MakeTypeAssertion(Response.prototype)
+webidl.is.Response = webidl.util.MakeTypeAssertion(Response)
 
 module.exports = {
   isNetworkError,