--- a/lib/web/fetch/webidl.js
+++ b/lib/web/fetch/webidl.js
@@ -12,6 +12,8 @@ const BIGINT = 6
 const NULL = 7
 const OBJECT = 8 // function and object
 
+const FunctionPrototypeSymbolHasInstance = Function.call.bind(Function.prototype[Symbol.hasInstance])
+
 /** @type {import('../../../types/webidl').Webidl} */
 const webidl = {
   converters: {},
@@ -45,7 +47,7 @@ webidl.errors.invalidArgument = function (context) {
 
 // https://webidl.spec.whatwg.org/#implements
 webidl.brandCheck = function (V, I) {
-  if (!I.prototype.isPrototypeOf(V)) { // eslint-disable-line no-prototype-builtins
+  if (!FunctionPrototypeSymbolHasInstance(I, V)) {
     const err = new TypeError('Illegal invocation')
     err.code = 'ERR_INVALID_THIS' // node compat.
     throw err
@@ -53,7 +55,7 @@ webidl.brandCheck = function (V, I) {
 }
 
 webidl.brandCheckMultiple = function (List) {
-  const prototypes = List.map((c) => webidl.util.MakeTypeAssertion(c.prototype))
+  const prototypes = List.map((c) => webidl.util.MakeTypeAssertion(c))
 
   return (V) => {
     if (prototypes.every(typeCheck => !typeCheck(V))) {
@@ -81,9 +83,8 @@ webidl.illegalConstructor = function () {
   })
 }
 
-const isPrototypeOf = Object.prototype.isPrototypeOf
-webidl.util.MakeTypeAssertion = function (Prototype) {
-  return (O) => isPrototypeOf.call(Prototype, O)
+webidl.util.MakeTypeAssertion = function (I) {
+  return (O) => FunctionPrototypeSymbolHasInstance(I, O)
 }
 
 // https://tc39.es/ecma262/#sec-ecmascript-data-types-and-values
@@ -462,13 +463,13 @@ webidl.nullableConverter = function (converter) {
   }
 }
 
-webidl.is.ReadableStream = webidl.util.MakeTypeAssertion(ReadableStream.prototype)
-webidl.is.Blob = webidl.util.MakeTypeAssertion(Blob.prototype)
-webidl.is.URLSearchParams = webidl.util.MakeTypeAssertion(URLSearchParams.prototype)
-webidl.is.File = webidl.util.MakeTypeAssertion((globalThis.File ?? require('node:buffer').File).prototype)
-webidl.is.URL = webidl.util.MakeTypeAssertion(URL.prototype)
-webidl.is.AbortSignal = webidl.util.MakeTypeAssertion(AbortSignal.prototype)
-webidl.is.MessagePort = webidl.util.MakeTypeAssertion(MessagePort.prototype)
+webidl.is.ReadableStream = webidl.util.MakeTypeAssertion(ReadableStream)
+webidl.is.Blob = webidl.util.MakeTypeAssertion(Blob)
+webidl.is.URLSearchParams = webidl.util.MakeTypeAssertion(URLSearchParams)
+webidl.is.File = webidl.util.MakeTypeAssertion(globalThis.File ?? require('node:buffer').File)
+webidl.is.URL = webidl.util.MakeTypeAssertion(URL)
+webidl.is.AbortSignal = webidl.util.MakeTypeAssertion(AbortSignal)
+webidl.is.MessagePort = webidl.util.MakeTypeAssertion(MessagePort)
 
 // https://webidl.spec.whatwg.org/#es-DOMString
 webidl.converters.DOMString = function (V, prefix, argument, opts) {