--- a/lib/web/websocket/stream/websocketerror.js
+++ b/lib/web/websocket/stream/websocketerror.js
@@ -78,6 +78,6 @@ Object.defineProperties(WebSocketError.prototype, {
   }
 })
 
-webidl.is.WebSocketError = webidl.util.MakeTypeAssertion(WebSocketError.prototype)
+webidl.is.WebSocketError = webidl.util.MakeTypeAssertion(WebSocketError)
 
 module.exports = { WebSocketError, createUnvalidatedWebSocketError }