https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Function/Symbol.hasInstance#reverting_to_default_instanceof_behavior

- main
```shell
benchmark              avg (min … max) p75   p99    (min … top 1%)
-------------------------------------- -------------------------------
webidl.is.FormData (ok)   7.17 ns/iter   6.91 ns  █▄
                 (6.20 ns … 150.63 ns)  14.65 ns ▁██▃▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁
webidl.is.FormData (b..   8.81 ns/iter   9.03 ns █
                 (7.62 ns … 101.73 ns)  18.87 ns █▄▁▄▁▁▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁
instanceof (ok)          22.41 ns/iter  23.14 ns  █
                 (19.60 ns … 63.67 ns)  43.04 ns ▁█▂▁▄▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁
instanceof (bad)         23.51 ns/iter  22.61 ns  █
                (20.53 ns … 122.12 ns)  47.97 ns ▁█▂▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁

                        ┌                                            ┐
webidl.is.FormData (ok) ┤■ 7.17 ns
webidl.is.FormData (b.. ┤■■■■ 8.81 ns
        instanceof (ok) ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 22.41 ns
       instanceof (bad) ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 23.51 ns
                        └                                            ┘
```

- this patch

```
benchmark              avg (min … max) p75   p99    (min … top 1%)
-------------------------------------- -------------------------------
webidl.is.FormData (ok)   3.69 ns/iter   3.42 ns  █
                 (3.20 ns … 139.82 ns)   7.25 ns ▁█▁▂▂▁▁▁▁▂▁▁▁▁▁▁▁▁▁▁▁
webidl.is.FormData (b.. 894.14 ps/iter 854.49 ps    █
                (756.84 ps … 47.92 ns)   1.29 ns ▁▁▇█▃▂▂▂▁▁▃▁▁▁▁▁▁▂▁▁▁
instanceof (ok)          23.62 ns/iter  24.00 ns   █
                (19.04 ns … 121.68 ns)  43.41 ns ▂▁█▄▇▁▁▂▁▁▂▁▁▁▁▁▁▁▁▁▁
instanceof (bad)         23.55 ns/iter  23.17 ns  █
                 (20.53 ns … 99.58 ns)  44.46 ns ▁█▂▁▃▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁

                        ┌                                            ┐
webidl.is.FormData (ok) ┤■■■■■ 3.69 ns
webidl.is.FormData (b.. ┤■ 894.14 ps
        instanceof (ok) ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 23.62 ns
       instanceof (bad) ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 23.55 ns
                        └                                            ┘
```