@tsctx 2024-10-12
Unfortunately, when examined, there is no difference in the real world, but it is always faster than the original isPrototypeOf.

```shell
import { bench, run } from "mitata";

const FunctionPrototypeSymbolHasInstance = Function.call.bind(
  Function.prototype[Symbol.hasInstance]
);

function A() {
  if (!(this instanceof A)) {
    return new A();
  }

  return this;
}

function B() {
  if (!(this instanceof B)) {
    return new B();
  }

  return this;
}

class State {
  _now = -1;
  now() {
    return (this._now = (this._now + 1) & 0x7fffffff);
  }
}

const array = [];

for (let i = 0; i < 0xff; ++i) {
  array.push(Math.random() > 0.5 ? new A() : new B());
}

const state = new State();

const isPrototypeOf = Object.prototype.isPrototypeOf;
const APrototype = A.prototype;

bench("a", () => array[state.now() & 0xff] instanceof A);
bench("b", () =>
  FunctionPrototypeSymbolHasInstance(A, array[state.now() & 0xff])
);
bench("c", () => isPrototypeOf.call(A.prototype, array[state.now() & 0xff]));
bench("c", () => isPrototypeOf.call(APrototype, array[state.now() & 0xff]));

await run();
```


```
benchmark              avg (min … max) p75   p99    (min … top 1%)
-------------------------------------- -------------------------------
a                         6.56 ns/iter   6.13 ns  █
                  (5.37 ns … 88.99 ns)  16.97 ns ▁█▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁
b                         8.02 ns/iter   7.98 ns  █
                 (6.79 ns … 194.97 ns)  18.97 ns ▁█▄▂▂▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁
c                        14.84 ns/iter  14.18 ns  █
                 (12.82 ns … 84.62 ns)  34.67 ns ▁█▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁
c                         8.46 ns/iter   8.01 ns  █
                  (7.25 ns … 82.47 ns)  21.26 ns ▁█▂▁▂▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁
```